+++ src/zinx/znet/Message.go
@@ -1,24 +1,24 @@
 package znet
 
 type Message struct {
-	Id      uint64
-	DataLen uint64
+	Id      uint32
+	DataLen uint32
 	Data    []byte
 }
 
-func NewMsg(id uint64,data []byte) *Message{
+func NewMsg(id uint32, data []byte) *Message {
 	return &Message{
 		Id:      id,
-		DataLen: uint64(len(data)),
+		DataLen: uint32(len(data)),
 		Data:    data,
 	}
 }
 
-func (m *Message) GetMsgID() uint64 {
+func (m *Message) GetMsgID() uint32 {
 	return m.Id
 }
 
-func (m *Message) GetMsgLen() uint64 {
+func (m *Message) GetMsgLen() uint32 {
 	return m.DataLen
 }
 
@@ -26,16 +26,14 @@ func (m *Message) GetMsgData() []byte {
 	return m.Data
 }
 
-func (m *Message) SetMsgID(id uint64) {
+func (m *Message) SetMsgID(id uint32) {
 	m.Id = id
 }
 
-func (m *Message) SetMsgLen(len uint64) {
+func (m *Message) SetMsgLen(len uint32) {
 	m.DataLen = len
 }
 
 func (m *Message) SetData(data []byte) {
 	m.Data = data
 }
-
-
