+++ src/zinx/znet/Request.go
@@ -19,6 +19,6 @@ func (r *Request) GetConn() ziface.IConn {
 func (r *Request) GetData() []byte {
 	return r.msg.GetMsgData()
 }
-func (r *Request) GetMsgID() uint64 {
+func (r *Request) GetMsgID() uint32 {
 	return r.msg.GetMsgID()
 }
